+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/specialized/BlobClientBase.java
@@ -30,7 +30,7 @@ import com.azure.storage.blob.BlobServiceVersion;
 import com.azure.storage.blob.implementation.AzureBlobStorageImpl;
 import com.azure.storage.blob.implementation.AzureBlobStorageImplBuilder;
 import com.azure.storage.blob.implementation.accesshelpers.BlobPropertiesConstructorProxy;
-import com.azure.storage.blob.implementation.models.BlobLayout;
+import com.azure.storage.blob.implementation.models.BlobLayoutInternal;
 import com.azure.storage.blob.implementation.models.BlobPropertiesInternalGetProperties;
 import com.azure.storage.blob.implementation.models.BlobTag;
 import com.azure.storage.blob.implementation.models.BlobTags;
@@ -1853,7 +1853,7 @@ public class BlobClientBase {
      * @return A response emitting all blob layout information.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedIterable<BlobLayoutInfo> getLayout(BlobGetLayoutOptions options) {
+    public PagedIterable<com.azure.storage.blob.models.BlobLayout> getLayout(BlobGetLayoutOptions options) {
         return getLayout(options, Context.NONE);
     }
 
@@ -1865,33 +1865,40 @@ public class BlobClientBase {
      * @return A response emitting all blob layout information.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedIterable<BlobLayoutInfo> getLayout(BlobGetLayoutOptions options, Context context) {
+    public PagedIterable<com.azure.storage.blob.models.BlobLayout> getLayout(BlobGetLayoutOptions options,
+        Context context) {
         Context finalContext = context == null ? Context.NONE : context;
         BlobGetLayoutOptions finalOptions = options == null ? new BlobGetLayoutOptions() : options;
 
-        BiFunction<String, Integer, PagedResponse<BlobLayoutInfo>> pageRetriever = (continuationToken, pageSize) -> {
-            BlobRange range = finalOptions.getRange() == null ? new BlobRange(0) : finalOptions.getRange();
-            BlobRequestConditions requestConditions = finalOptions.getRequestConditions() == null
-                ? new BlobRequestConditions()
-                : finalOptions.getRequestConditions();
-            Integer finalPageSize = pageSize == null ? finalOptions.getMaxResultsPerPage() : pageSize;
+        BiFunction<String, Integer, PagedResponse<com.azure.storage.blob.models.BlobLayout>> pageRetriever
+            = (continuationToken, pageSize) -> {
+                BlobRange range = finalOptions.getRange() == null ? new BlobRange(0) : finalOptions.getRange();
+                BlobRequestConditions requestConditions = finalOptions.getRequestConditions() == null
+                    ? new BlobRequestConditions()
+                    : finalOptions.getRequestConditions();
+                Integer finalPageSize = pageSize == null ? finalOptions.getMaxResultsPerPage() : pageSize;
 
-            Callable<ResponseBase<BlobsGetLayoutHeaders, BlobLayout>> operation = () -> this.azureBlobStorage.getBlobs()
-                .getLayoutWithResponse(containerName, blobName, snapshot, versionId, continuationToken, finalPageSize,
-                    null, range.toHeaderValue(), requestConditions.getLeaseId(), requestConditions.getTagsConditions(),
-                    requestConditions.getIfModifiedSince(), requestConditions.getIfUnmodifiedSince(),
-                    requestConditions.getIfMatch(), requestConditions.getIfNoneMatch(), null, customerProvidedKey,
-                    finalContext);
+                Callable<ResponseBase<BlobsGetLayoutHeaders, BlobLayoutInternal>> operation
+                    = () -> this.azureBlobStorage.getBlobs()
+                        .getLayoutWithResponse(containerName, blobName, snapshot, versionId, continuationToken,
+                            finalPageSize, null, range.toHeaderValue(), requestConditions.getLeaseId(),
+                            requestConditions.getTagsConditions(), requestConditions.getIfModifiedSince(),
+                            requestConditions.getIfUnmodifiedSince(), requestConditions.getIfMatch(),
+                            requestConditions.getIfNoneMatch(), null, customerProvidedKey, finalContext);
 
-            ResponseBase<BlobsGetLayoutHeaders, BlobLayout> response
-                = sendRequest(operation, null, BlobStorageException.class);
-            BlobLayoutInfo value = ModelHelper.transformBlobLayoutInfo(response);
-            BlobLayout layout = response.getValue();
+                ResponseBase<BlobsGetLayoutHeaders, BlobLayoutInternal> response
+                    = sendRequest(operation, null, BlobStorageException.class);
+                BlobLayoutInfo value = ModelHelper.transformBlobLayoutInfo(response);
+                com.azure.storage.blob.models.BlobLayout publicValue = value == null
+                    ? null
+                    : new com.azure.storage.blob.models.BlobLayout(value.getRanges(), null,
+                        response.getValue() == null ? null : response.getValue().getNextMarker(), finalPageSize, value);
+                BlobLayoutInternal layout = response.getValue();
 
-            return new PagedResponseBase<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
-                value == null ? Collections.emptyList() : Collections.singletonList(value),
-                layout == null ? null : layout.getNextMarker(), response.getDeserializedHeaders());
-        };
+                return new PagedResponseBase<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
+                    publicValue == null ? Collections.emptyList() : Collections.singletonList(publicValue),
+                    layout == null ? null : layout.getNextMarker(), response.getDeserializedHeaders());
+            };
 
         return new PagedIterable<>(pageSize -> pageRetriever.apply(null, pageSize), pageRetriever);
     }